fix(arch): resolve -Wconversion warnings in RegVM emitter and VM
【动机】
MSVC 的 `/WX`（警告视为错误）将 C4244（类型转换可能丢失数据）和 C4100（未引用参数）等警告视为编译错误，阻碍了 RegVM 的集成。需要消除这些警告使项目能正常通过编译。

【原理】
1. Emitter.cpp 中将 `size_t` 与 `int` 之间的隐式转换改为显式 `static_cast<int>`，消除 C4244 警告。
2. VM.cpp 中对未使用的局部变量 `ax` 添加 `(void)ax;` 语句，消除 C4100 未引用变量警告。
3. VM.cpp 中调用 `execSliceGet` 和 `execSliceSet` 时，将 `int c` 显式转换为 `uint8_t`，避免截断警告。

+++ src/regvm/compiler/Emitter.cpp
@@ -657,7 +657,7 @@ int Emitter::emit(IRGraph* graph, Chunk& chunk) {
                           inst.node->op == IROp::BuildMatrix || inst.node->op == IROp::IndexGet ||
                           inst.node->op == IROp::IndexSet || inst.node->op == IROp::SliceGet ||
                           inst.node->op == IROp::SliceSet)) {
-            int argsCount = inst.node->dataInputs.size();
+            int argsCount = static_cast<int>(inst.node->dataInputs.size());
             if (dynamicSpillBase + argsCount > absoluteMaxReg) {
                 absoluteMaxReg = dynamicSpillBase + argsCount;
             }
